feat: add icon for Meta's Threads.net Platform (#55)
* Added Icons for Meta's Threads.Net

Threads is an American social media and social networking service owned and operated by Meta Platforms. It operates in a similar manner to other microblogging platforms such as Twitter. Users have the ability to post and share text, images, and videos, as well as interact with other users' posts through replies, reposts, and likes. Threads is closely integrated with Instagram, requiring users to have an Instagram account and use the same handle. The application is available on iOS and Android devices with full functionality, while the web version offers limited functionality.

* SVG Class Removed

* Removed BW Threads Logo

* Deleted Threads Color Svg

* Added Threads Plain SVG